Правки

+++ Task_Manegr/MetricsAgent/AllHddMetricsResponse.cs
@@ -10,7 +10,7 @@ namespace MetricsAgent.Controllers
     public class HddMetricDto
     {
         public DateTimeOffset Time { get; set; }
-        public int Value { get; set; }
+        public long Value { get; set; }
         public int Id { get; set; }
     }
